JPAK 2.0 Refactor
*	Changed the Github URLs in ALL files to the correct one under TeskeVirtualSystem github
*	Separated the class files in jssrc folder
*	Made a grunt task to compile it into one file
*	Adapted extpacker to the changes
