+++ .bowerrc
@@ -1,0 +1,4 @@
+{
+  "directory": "3rdparty",
+  "json": "bower.json"
+}
