+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules
+3rdparty
+dist
