+++ doc/JPAK1.0.txt
@@ -6,7 +6,7 @@
                                          
 Multiuse Javascript Package 
 By: Lucas Teske
-https://github.com/racerxdl/jpak
+https://github.com/TeskeVirtualSystem/jpak
 
 JPAK v1.0 Specification File
