+++ doc/JPAK2/Extended Mode/README.md
@@ -12,7 +12,7 @@
                                                
 Multiuse Javascript Package 
 By: Lucas Teske
-https://github.com/racerxdl/jpak
+https://github.com/TeskeVirtualSystem/jpak
 
 #JPAK v2.0 Extended Mode
