+++ doc/JPAK2/README.md
@@ -6,7 +6,7 @@
                                                                                        
 Multiuse Javascript Package 
 By: Lucas Teske
-https://github.com/racerxdl/jpak
+https://github.com/TeskeVirtualSystem/jpak
 
 #JPAK v2.0 Specification File
