+++ jsdoc/symbols/src/jpak.js.html
@@ -15,7 +15,7 @@
 <span class='line'>  8</span>                                          
 <span class='line'>  9</span> Multiuse Javascript Package 
 <span class='line'> 10</span> By: Lucas Teske
-<span class='line'> 11</span> https://github.com/racerxdl/jpak
+<span class='line'> 11</span> https://github.com/TeskeVirtualSystem/jpak
 <span class='line'> 12</span> */</span><span class="WHIT">
 <span class='line'> 13</span> 
 <span class='line'> 14</span> </span><span class="COMM">/**
