+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "JPAK",
+  "version": "2.0.0",
+  "repository" : {
+    "type" : "git",
+    "url" : "https://github.com/TeskeVirtualSystem/jpak"
+  },
+  "dependencies" : {
+  },
+  "keywords" : [
+    "file system",
+    "distributed",
+    "package"
+  ],
+  "devDependencies": {
+    "grunt": "~0.4.5",
+    "grunt-cli": "~0.1.13",
+    "grunt-html2js": "~0.2.7",
+    "grunt-contrib-uglify": "~0.5.0",
+    "grunt-contrib-less": "~0.11.2",
+    "grunt-contrib-clean": "~0.5.0",
+    "grunt-contrib-connect": "~0.8.0",
+    "grunt-contrib-compress": "~0.9.1",
+    "grunt-contrib-concat": "~0.4.0",
+    "grunt-contrib-jshint": "~0.10.0",
+    "grunt-contrib-watch": "~0.6.1",
+    "grunt-bower-task": "~0.3.4"
+  },
+  "license": "MIT"
+}
