+++ tools/JPAK Explorer Win/JPAKTool/FileEntry.cs
@@ -7,7 +7,7 @@
  *                                            *
  *Multiuse Javascript Package                 *
  *By: Lucas Teske                             *
- *https://github.com/racerxdl/jpak            *
+ *https://github.com/TeskeVirtualSystem/jpak  *
  **********************************************/
 using System;
 using System.Collections.Generic;
