+++ tools/jpaktool.py
@@ -9,7 +9,7 @@
                                                 
 Multiuse Javascript Package 
 By: Lucas Teske
-https://github.com/racerxdl/jpak
+https://github.com/TeskeVirtualSystem/jpak
 
 '''
 import struct, os, json, sys, zlib, md5, base64
