+++ tools/unpacker.py
@@ -9,7 +9,7 @@ print '''
                                      |_|                                                                     
 Multiuse Javascript Package 
 By: Lucas Teske
-https://github.com/racerxdl/jpak
+https://github.com/TeskeVirtualSystem/jpak
 '''
 
 import struct, os, json, sys
